requirements.txt
psycopg2-binary==2.8.6
pyautogui==0.9.50
pytesseract==0.3.6
python-Levenshtein==0.12.0
requests==2.25.0
